requirements.txt
flask
flask-bootstrap
requests
faiss-cpu
pandas
numpy
tensorflow-text
tensorflow-hub
